Split files out of index.ts
